create FileList component instead of FileItem
