+++ demo/site/src/common/components/form/FileList.module.scss
@@ -1,3 +1,12 @@
+.fileList {
+    list-style: none;
+    margin: 0;
+    padding: 0;
+    display: flex;
+    flex-direction: column;
+    gap: var(--spacing-s100);
+}
+
 .fileItem {
     display: flex;
     align-items: center;
