+++ demo/site/src/common/components/form/FileUploadField.module.scss
@@ -15,12 +15,3 @@
     white-space: nowrap;
     border: 0;
 }
-
-.fileList {
-    list-style: none;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    flex-direction: column;
-    gap: var(--spacing-s100);
-}
